requirements.txt
discord.py
aiohttp
websockets
chardet
redis
fuzzywuzzy